requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-text-splitters
langchain
langchain-community
faiss-cpu
pypdf
langchain-google-genai 